+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 8.1.0
+
+- updated phpunit testing
+- updated example public index.php
+
 ## 8.0.1
 
 - extracted markup rendering from bundling